feat: change footer items
