+++ src/components/molecules/ChoiceCard/ChoiceCard.tsx
@@ -1,5 +1,5 @@
 import { Checkbox } from '../../atoms/Checkbox/Checkbox';
-import { Picture } from '../../atoms/Picture/Picture';
+// import { Picture } from '../../atoms/Picture/Picture';
 import { Text } from '../../atoms/Text/Text';
 
 import S from './ChoiceCard.module.scss';
@@ -7,7 +7,7 @@ import S from './ChoiceCard.module.scss';
 export const ChoiceCard = () => {
   return (
     <li className={S.card}>
-      <Picture size='small' image='choice' />
+      {/* <Picture size='small' image='choice' /> */}
       <Checkbox>
         <Text color='grey' size='default'>
           sneakers
