+++ src/components/molecules/HeaderFooterMenu/HeaderFooterMenu.tsx
@@ -2,52 +2,16 @@ import { HeaderFooterLink } from '../../atoms/HeaderFooterLink/HeaderFooterLink'
 import S from './HeaderFooterMenu.module.scss';
 import useIsStaffPage from '../../../hooks/useIsStaffPage';
 import { useCallback } from 'react';
+import { toAnchor } from './utils';
+import { HeaderFooterType } from '../../../utils/types';
+import { footerItems, headerItems } from './constants';
 
-const items = [
-  {
-    name: 'Catalog',
-    link: '#catalog',
-  },
-  {
-    name: 'About us',
-    link: '#about',
-  },
-  {
-    name: 'Product selection',
-    link: '#selection',
-  },
-  {
-    name: 'Our team',
-    link: '#team',
-  },
-  {
-    name: 'FAQ',
-    link: '#faq',
-  },
-  {
-    name: 'For staff',
-    link: '/staff',
-  },
-];
+export const HeaderFooterMenu = ({ type }: HeaderFooterType) => {
+  const items = type === 'header' ? headerItems : footerItems;
 
-export const HeaderFooterMenu = () => {
   const isStaffPage = useIsStaffPage();
 
-  const handleLinkClick = useCallback(
-    (event: React.MouseEvent<HTMLAnchorElement>) => {
-      event.preventDefault();
-      const href = event.currentTarget.getAttribute('href');
-      const id = href?.substring(href.indexOf('#')) ?? '';
-
-      if (id.startsWith('#') && id.length > 1) {
-        const element = document.querySelector(id);
-        if (element) {
-          element.scrollIntoView({ behavior: 'smooth' });
-        }
-      }
-    },
-    []
-  );
+  const handleLinkClick = useCallback(toAnchor, []);
 
   const baseContent = items.map(({ name, link }, index) => (
     <li key={index}>
