+++ src/components/molecules/HeaderFooterMenu/utils.ts
@@ -1,0 +1,12 @@
+export const toAnchor = (event: React.MouseEvent<HTMLAnchorElement>) => {
+  event.preventDefault();
+  const href = event.currentTarget.getAttribute('href');
+  const id = href?.substring(href.indexOf('#')) ?? '';
+
+  if (id.startsWith('#') && id.length > 1) {
+    const element = document.querySelector(id);
+    if (element) {
+      element.scrollIntoView({ behavior: 'smooth' });
+    }
+  }
+};
