+++ src/components/organisms/HeaderFooter/HeaderFooter.tsx
@@ -1,9 +1,10 @@
 import useIsStaffPage from '../../../hooks/useIsStaffPage';
+import { HeaderFooterType } from '../../../utils/types';
 import { Logo } from '../../atoms/Logo/Logo';
 import { HeaderFooterMenu } from '../../molecules/HeaderFooterMenu/HeaderFooterMenu';
 import S from './HeaderFooter.module.scss';
 
-export const HeaderFooter = ({ type }: { type: 'header' | 'footer' }) => {
+export const HeaderFooter = ({ type }: HeaderFooterType) => {
   const isStaffPage = useIsStaffPage();
 
   const staffStyle = `${S.header} ${S.header_staff}`;
@@ -13,7 +14,7 @@ export const HeaderFooter = ({ type }: { type: 'header' | 'footer' }) => {
     <>
       <Logo color='white' size='default' link />
       <div className={S.container}>
-        <HeaderFooterMenu />
+        <HeaderFooterMenu type={type} />
         {/* здесь могла бы быть корзина :( */}
         {/* {type === 'header' && <HeaderCart />} */}
       </div>
