+++ src/components/organisms/ProductSelection/ProductSelection.module.scss
@@ -23,8 +23,9 @@
 
 .list {
   display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-gap: 2rem 7rem;
+  grid-template-columns: repeat(4, 1fr);
+  justify-content: space-between;
+  grid-gap: 2.8rem 7rem;
 }
 
 @media (max-width: 768px) {
