+++ src/utils/types.ts
@@ -1,0 +1,1 @@
+export type HeaderFooterType = { type: 'header' | 'footer' };
